Added config file for P1022DS
This patch added the config file for P1022DS.

The new config file contains the configurations is used to configure the L2
array. It enables L2 array and determines the entire L2 array is a single an
SRAM(256-Kbytes). The L2 memory-mapped SRAM base address is 0xf8f80000.

The on-chip boot ROM code uses the information to configure the L2 array and
copies a U-Boot image to the target address in the L2 SRAM, the target address
is 0xf8f80000.
After that, the e500 core jumps to the excution starting address in the L2
SRAM and starts to execute the code from the L2 SRAM, the excution starting
address is 0xf8f80000.

For details onconfiguration information, refer to the product manuals.

Signed-off-by: Ying Zhang <b40530@freescale.com>
Change-Id: Ifcb565f18eb057e4068099f06209e373a9783ec1
Reviewed-on: http://git.am.freescale.net:8181/1712
Reviewed-by: Schmitt Richard-B43082 <B43082@freescale.com>
Tested-by: Schmitt Richard-B43082 <B43082@freescale.com>