p1010rdb: Add config file
The new config file contains the configurations is used to configure the L2
array. It enables L2 array and determines the entire L2 array is a single an
SRAM(256-Kbytes). The L2 memory-mapped SRAM base address is 0xd0000000.

The on-chip boot ROM code uses the information to configure the L2 array and
copies a U-Boot image to the target address in the L2 SRAM, the target address
is 0xd0000000.
After that, the e500 core jumps to the excution starting address in the L2
SRAM and starts to execute the code from the L2 SRAM, the execution starting
address is 0xd0000000.

Signed-off-by: Ying Zhang <b40530@freescale.com>
Change-Id: I3bb40830ee4590a76f12ef33beddb4fd62a1f9cf
Reviewed-on: http://git.am.freescale.net:8181/6602
Reviewed-by: Emilian Medve <Emilian.Medve@freescale.com>
Tested-by: Emilian Medve <Emilian.Medve@freescale.com>
Reviewed-by: Xiaobo Xie <X.Xie@freescale.com>
Reviewed-by: Richard Schmitt <B43082@freescale.com>